fix item crash and fix crash when by item

+++ src/Zepheus.Database/Storage/EquipInfo.cs
@@ -9,11 +9,11 @@ namespace Zepheus.Database.Storage
         public short Slot { get; set; }
         public int EquipID { get; set; }
         public byte Upgrades { get; set; }
-        public byte IncStr { get; set; }
-        public byte IncEnd { get; set; }
-        public byte IncDex { get; set; }
-        public byte IncInt { get; set; }
-        public byte IncSpr { get; set; }
+        public ushort IncStr { get; set; }
+        public ushort IncEnd { get; set; }
+        public ushort IncDex { get; set; }
+        public ushort IncInt { get; set; }
+        public ushort IncSpr { get; set; }
         public DateTime? Expires { get; set; }
         public Character Character { get; set; }
